+++ codec/binarycodec/types/amount.go
@@ -214,7 +214,6 @@ func (a *Amount) ToJSON(p interfaces.BinaryParser, _ ...int) (any, error) {
 }
 
 func deserializeToken(data []byte) (map[string]any, error) {
-
 	var value string
 	var err error
 	if bytes.Equal(data[0:8], []byte{0x80, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00}) {
@@ -347,7 +346,6 @@ func deserializeMPTAmount(data []byte) (map[string]any, error) {
 // verifyXrpValue validates the format of an XRP amount value.
 // XRP values should not contain a decimal point because they are represented as integers as drops.
 func verifyXrpValue(value string) error {
-
 	r := regexp.MustCompile(`\d+`) // regex to match only digits
 	m := r.FindAllString(value, -1)
 
@@ -438,7 +436,6 @@ func verifyMPTValue(value string) error {
 
 // serializeXrpAmount serializes an XRP amount value.
 func serializeXrpAmount(value string) ([]byte, error) {
-
 	if verifyXrpValue(value) != nil {
 		return nil, verifyXrpValue(value)
 	}
@@ -452,7 +449,7 @@ func serializeXrpAmount(value string) ([]byte, error) {
 	valWithPosBit := val | PosSignBitMask
 	valBytes := make([]byte, NativeAmountByteLength)
 
-	binary.BigEndian.PutUint64(valBytes, uint64(valWithPosBit))
+	binary.BigEndian.PutUint64(valBytes, valWithPosBit)
 
 	return valBytes, nil
 }
@@ -467,7 +464,6 @@ func serializeXrpAmount(value string) ([]byte, error) {
 
 // SerializeIssuedCurrencyValue serializes the value field of an issued currency amount to its bytes representation.
 func SerializeIssuedCurrencyValue(value string) ([]byte, error) {
-
 	if verifyIOUValue(value) != nil {
 		return nil, verifyIOUValue(value)
 	}
@@ -525,7 +521,7 @@ func SerializeIssuedCurrencyValue(value string) ([]byte, error) {
 	// TODO: Check if this is still needed
 	//nolint:gosec // G115: Potential hardcoded credentials (gosec)
 	serial |= (uint64(exp+97) << 54) // if the exponent is positive, set the exponent bits to the exponent + 97
-	serial |= uint64(mantissa)       // last 54 bits are mantissa
+	serial |= mantissa               // last 54 bits are mantissa
 
 	serialReturn := make([]byte, 8)
 	binary.BigEndian.PutUint64(serialReturn, serial)
@@ -536,7 +532,6 @@ func SerializeIssuedCurrencyValue(value string) ([]byte, error) {
 // serializeIssuedCurrencyCode serializes an issued currency code to its bytes representation.
 // The currency code can be 3 allowed string characters, or 20 bytes of hex.
 func serializeIssuedCurrencyCode(currency string) ([]byte, error) {
-
 	currency = strings.TrimPrefix(currency, "0x")                                    // remove the 0x prefix if it exists
 	if currency == "XRP" || currency == "0000000000000000000000005852500000000000" { // if the currency code is uppercase XRP, return an error
 		return nil, &InvalidCodeError{Disallowed: "XRP uppercase"}
@@ -550,7 +545,6 @@ func serializeIssuedCurrencyCode(currency string) ([]byte, error) {
 	}
 
 	return nil, &InvalidCodeError{Disallowed: currency}
-
 }
 
 func serializeIssuedCurrencyCodeHex(currency string) ([]byte, error) {
@@ -561,7 +555,6 @@ func serializeIssuedCurrencyCodeHex(currency string) ([]byte, error) {
 	}
 
 	if bytes.HasPrefix(decodedHex, []byte{0x00}) {
-
 		if bytes.Equal(decodedHex[12:15], []byte{0x00, 0x00, 0x00}) {
 			return make([]byte, 20), nil
 		}
@@ -570,13 +563,11 @@ func serializeIssuedCurrencyCodeHex(currency string) ([]byte, error) {
 			return nil, errInvalidCurrencyCode
 		}
 		return decodedHex, nil
-
 	}
 	return decodedHex, nil
 }
 
 func serializeIssuedCurrencyCodeChars(currency string) ([]byte, error) {
-
 	r := regexp.MustCompile(IOUCodeRegex) // regex to check if the currency code is valid
 	m := r.FindAllString(currency, -1)
 
@@ -594,7 +585,6 @@ func serializeIssuedCurrencyCodeChars(currency string) ([]byte, error) {
 // The currency code can be 3 allowed string characters, or 20 bytes of hex in standard currency format (e.g. with "00" prefix)
 // or non-standard currency format (e.g. without "00" prefix)
 func serializeIssuedCurrencyAmount(value, currency, issuer string) ([]byte, error) {
-
 	var valBytes []byte
 	var err error
 	if value == "0" {
@@ -691,7 +681,6 @@ func isPositive(value byte) bool {
 }
 
 func containsInvalidIOUCodeCharactersHex(currency []byte) bool {
-
 	r := regexp.MustCompile(IOUCodeRegex) // regex to check if the currency code is valid
 	m := r.FindAll(currency, -1)
