+++ internal/tx/nftoken/nftoken_offer.go
@@ -210,7 +210,7 @@ func tokenOfferCreateApply(
 	offerNode := tokenDirResult.Page
 
 	// Serialize the offer
-	flags := uint32(NFTokenCreateOfferFlagSellNFToken) // Always a sell offer
+	flags := NFTokenCreateOfferFlagSellNFToken // Always a sell offer
 
 	offerData, err := serializeNFTokenOfferRaw(
 		accountID, tokenID,
